refactor: extract verifications table
