+++ client/src/modules/AutoTest/components/index.tsx
@@ -1,3 +1,4 @@
 export { default as TaskCard } from './TaskCard/TaskCard';
 export { default as TaskCardColumn } from './TaskCardColumn/TaskCardColumn';
 export { default as TaskDeadlineDate } from './TaskDeadlineDate/TaskDeadlineDate';
+export { default as VerificationsTable } from './VerificationsTable/VerificationsTable';
